Change how we handle empty vector comparisons
We used to return 0 (iow. max distance) for empty vectors, but this is
based on historical reasons where we needed to handle the empty case as
a valid state where you were in agreement with all parties.

We don't have that need anymore, but we have the need to distinguish
unknowable distance from max distance. So we change how empty
comparisons are handeled to return `null` instead of `0`.